Rename the appModule to make it importable and provide an alias to map it into right binary.
